requirements.txt
Django==4.2.4
django-restframework==0.0.1
djangorestframework==3.14.0
environs==9.5.0
Pillow==10.0.0
psycopg2==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
